swml/parsing/inlines-2.dat


#data
aa[B[bbb]]o
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <b>
|         "bbb"
|       "o"

#data
aa[I[bbb]]o
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <i>
|         "bbb"
|       "o"

#data
aa[U[bbb]]o
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <u>
|         "bbb"
|       "o"

#data
aa[ASIS[bbb]]o
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <sw asis>
|         "bbb"
|       "o"

#data
aa[ASIS[]]o
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <sw asis>
|       "o"

#data
aa[EMPH[bbb]]o
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <sw emph>
|         "bbb"
|       "o"

#data
aa[EMPH(a d)[bbb]]o
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <sw emph>
|         class="a d"
|         "bbb"
|       "o"

#data
aa[SNIP[bbb]]o
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <sw snip>
|         "bbb"
|       "o"

#data
aa[SNIP[]]o
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <sw snip>
|       "o"

#data
aa[DOTABOVE[bbb]]o
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <sw dotabove>
|         "bbb"
|       "o"

#data
aa[VECTOR[bbb]]o
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <sw vector>
|         "bbb"
|       "o"

#data
aa[YOKO[bbb]]o
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <sw yoko>
|         "bbb"
|       "o"

#data
aa[SMALLCAPS[bbb]]o
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <sw smallcaps>
|         "bbb"
|       "o"

#data
aa[TATE[bbb]]o
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <sw sw-tate>
|         "bbb"
|       "o"

#data
aa[L[bbb]]o
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <sw sw-l>
|         "bbb"
|       "o"

#data
aa[R[bbb]]o
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <sw sw-r>
|         "bbb"
|       "o"

#data
aa[V[bbb]]o
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <sw sw-v>
|         "bbb"
|       "o"
